# frontend/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FC Online Draft System",
    "description": "Realtime multi-team player draft system for Vietnamese FC Online pro tournament.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/tournaments": {
      "get": {
        "tags": [
          "Tournaments"
        ],
        "summary": "List Tournaments",
        "description": "List all tournaments (newest first).",
        "operationId": "list_tournaments_api_tournaments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TournamentListItem"
                  },
                  "type": "array",
                  "title": "Response List Tournaments Api Tournaments Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tournaments"
        ],
        "summary": "Create Tournament",
        "description": "Create a new tournament with validated TournamentRules (ADMIN only).",
        "operationId": "create_tournament_api_tournaments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTournamentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{tournament_id}": {
      "get": {
        "tags": [
          "Tournaments"
        ],
        "summary": "Get Tournament",
        "description": "Get tournament details with teams.",
        "operationId": "get_tournament_api_tournaments__tournament_id__get",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tournament Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{tournament_id}/complete": {
      "post": {
        "tags": [
          "Tournaments"
        ],
        "summary": "Complete Tournament",
        "description": "Manually mark a tournament as COMPLETED (ADMIN only).",
        "operationId": "complete_tournament_api_tournaments__tournament_id__complete_post",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tournament Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tournaments/{tournament_id}/matches": {
      "post": {
        "tags": [
          "Matches"
        ],
        "summary": "Create Match",
        "description": "Schedule a match between two teams in a tournament (ADMIN only).",
        "operationId": "create_match_api_tournaments__tournament_id__matches_post",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tournament Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMatchRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Matches"
        ],
        "summary": "List Tournament Matches",
        "description": "List all scheduled and played matches for a tournament.",
        "operationId": "list_tournament_matches_api_tournaments__tournament_id__matches_get",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tournament Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MatchDetailResponse"
                  },
                  "title": "Response List Tournament Matches Api Tournaments  Tournament Id  Matches Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/matches/{match_id}": {
      "get": {
        "tags": [
          "Matches"
        ],
        "summary": "Get Match Detail",
        "description": "Get match details with SIMULTANEOUS ban secrecy applied (BR-M04).",
        "operationId": "get_match_detail_api_matches__match_id__get",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Match Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/matches/{match_id}/bans/start": {
      "post": {
        "tags": [
          "Matches"
        ],
        "summary": "Start Ban Phase",
        "description": "Start Tactical Ban Phase for a SCHEDULED match (ADMIN only).",
        "operationId": "start_ban_phase_api_matches__match_id__bans_start_post",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Match Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/matches/{match_id}/bans": {
      "post": {
        "tags": [
          "Matches"
        ],
        "summary": "Submit Ban",
        "description": "Submit a single player card ban for the authenticated team.",
        "operationId": "submit_ban_api_matches__match_id__bans_post",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Match Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitBanRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BanItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/matches/{match_id}/bans/{ban_id}": {
      "delete": {
        "tags": [
          "Matches"
        ],
        "summary": "Delete Ban",
        "description": "Delete a pending ban before confirmation.",
        "operationId": "delete_ban_api_matches__match_id__bans__ban_id__delete",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Match Id"
            }
          },
          {
            "name": "ban_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ban Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/matches/{match_id}/bans/confirm": {
      "post": {
        "tags": [
          "Matches"
        ],
        "summary": "Confirm Bans",
        "description": "Confirm bans for the authenticated team.",
        "operationId": "confirm_bans_api_matches__match_id__bans_confirm_post",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Match Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/matches/{match_id}/complete": {
      "post": {
        "tags": [
          "Matches"
        ],
        "summary": "Complete Match",
        "description": "Mark match as COMPLETED (ADMIN only).",
        "operationId": "complete_match_api_matches__match_id__complete_post",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Match Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BanItemResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "banningTeamId": {
            "type": "string",
            "format": "uuid",
            "title": "Banningteamid"
          },
          "targetTeamId": {
            "type": "string",
            "format": "uuid",
            "title": "Targetteamid"
          },
          "playerSeasonId": {
            "type": "string",
            "format": "uuid",
            "title": "Playerseasonid"
          },
          "playerName": {
            "type": "string",
            "title": "Playername"
          },
          "position": {
            "type": "string",
            "title": "Position"
          },
          "rating": {
            "type": "integer",
            "title": "Rating",
            "default": 0
          },
          "seasonCode": {
            "type": "string",
            "title": "Seasoncode"
          },
          "seasonBadgeUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seasonbadgeurl"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "banningTeamId",
          "targetTeamId",
          "playerSeasonId",
          "playerName",
          "position",
          "seasonCode",
          "createdAt"
        ],
        "title": "BanItemResponse"
      },
...
      "CreateMatchRequest": {
        "properties": {
          "homeTeamId": {
            "type": "string",
            "format": "uuid",
            "title": "Hometeamid"
          },
          "awayTeamId": {
            "type": "string",
            "format": "uuid",
            "title": "Awayteamid"
          },
          "scheduledAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduledat"
          }
        },
        "type": "object",
        "required": [
          "homeTeamId",
          "awayTeamId"
        ],
        "title": "CreateMatchRequest"
      },
...
      "CreateTournamentRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "rules": {
            "$ref": "#/components/schemas/TournamentRulesSchema"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateTournamentRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MatchDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tournamentId": {
            "type": "string",
            "format": "uuid",
            "title": "Tournamentid"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "scheduledAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduledat"
          },
          "banStartedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Banstartedat"
          },
          "banExpiresAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Banexpiresat"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "serverTime": {
            "type": "string",
            "title": "Servertime"
          },
          "rulesSnapshot": {
            "additionalProperties": true,
            "type": "object",
            "title": "Rulessnapshot"
          },
          "homeTeam": {
            "$ref": "#/components/schemas/app__match__api__TeamSummary"
          },
          "awayTeam": {
            "$ref": "#/components/schemas/app__match__api__TeamSummary"
          },
          "bans": {
            "items": {
              "$ref": "#/components/schemas/BanItemResponse"
            },
            "type": "array",
            "title": "Bans",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "tournamentId",
          "status",
          "version",
          "serverTime",
          "rulesSnapshot",
          "homeTeam",
          "awayTeam"
        ],
        "title": "MatchDetailResponse"
      },
...
      "SubmitBanRequest": {
        "properties": {
          "playerSeasonId": {
            "type": "string",
            "format": "uuid",
            "title": "Playerseasonid"
          },
          "targetTeamId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetteamid"
          }
        },
        "type": "object",
        "required": [
          "playerSeasonId"
        ],
        "title": "SubmitBanRequest"
      },
...
      "TournamentListItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "rules": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TournamentRulesSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "teamCount": {
            "type": "integer",
            "title": "Teamcount"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "teamCount",
          "createdAt"
        ],
        "title": "TournamentListItem"
      },
      "TournamentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "rules": {
            "$ref": "#/components/schemas/TournamentRulesSchema"
          },
          "teams": {
            "items": {
              "$ref": "#/components/schemas/app__tournament__api__TeamSummary"
            },
            "type": "array",
            "title": "Teams"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "rules",
          "teams",
          "createdAt",
          "updatedAt"
        ],
        "title": "TournamentResponse"
      },
      "TournamentRulesSchema": {
        "properties": {
          "rulesVersion": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Rulesversion",
            "default": 1
          },
          "rosterSize": {
            "type": "integer",
            "maximum": 60.0,
            "minimum": 1.0,
            "title": "Rostersize",
            "default": 24
          },
          "budget": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Budget",
            "default": 305
          },
          "pickTimeSeconds": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Picktimeseconds",
            "default": 30
          },
          "uniqueBy": {
            "type": "string",
            "enum": [
              "PLAYER",
              "CARD"
            ],
            "title": "Uniqueby",
            "default": "PLAYER"
          },
          "timeoutPolicy": {
            "type": "string",
            "enum": [
              "SKIP_TURN",
              "AUTO_PICK_CHEAPEST"
            ],
            "title": "Timeoutpolicy",
            "default": "AUTO_PICK_CHEAPEST"
          },
          "allowedSeasonIds": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Allowedseasonids"
          },
          "banCount": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Bancount",
            "default": 5
          },
          "banTimeSeconds": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Bantimeseconds",
            "default": 60
          },
          "banTarget": {
            "type": "string",
            "enum": [
              "OPPONENT_ROSTER",
              "OWN_ROSTER"
            ],
            "title": "Bantarget",
            "default": "OPPONENT_ROSTER"
          },
          "banOrder": {
            "type": "string",
            "enum": [
              "SIMULTANEOUS",
              "ALTERNATING"
            ],
            "title": "Banorder",
            "default": "SIMULTANEOUS"
          }
        },
        "type": "object",
        "title": "TournamentRulesSchema",
        "description": "API-facing rules schema. Accepts camelCase or snake_case input, outputs camelCase."
      },
...
    }
  }
}